+++ demos/world/src/main/java/is/codion/demos/world/ui/CountryEditPanel.java
@@ -126,8 +126,7 @@ final class CountryEditPanel extends EntityEditPanel {
 		addInputPanel(Country.GOVERNMENTFORM);
 		addInputPanel(Country.HEADOFSTATE);
 		add(borderLayoutPanel()
-						.northComponent(Components.label()
-										.text("Avg. city population")
+						.northComponent(Components.label("Avg. city population")
 										.horizontalAlignment(SwingConstants.CENTER))
 						.centerComponent(averageCityPopulationField));
 	}